Refactor: Moved xr_core modules to protocol.py, and made protocol independent of other modules.
